added auth checking, delete trip endpoint

+++ routes/api/v1/controllers/trips.js
@@ -14,16 +14,21 @@ router.get('/', async function(req, res, next) {
 router.post('/add', async function(req, res, next){
     let session = req.session
     try {
-        console.log("poop1")
-        const newTrip = new req.models.Trip({
-            PrimaryUserEmail : session.account.username,
-            ShoppingList: [],
-            Users : [[session.account.username, session.account.name]]
-        })
-        console.log(newTrip)
-        await newTrip.save()
-        res.json({status:'success', ID: newTrip._id});
-        console.log("poop2")
+        if(req.session.isAuthenticated == true){
+            const newTrip = new req.models.Trip({
+                PrimaryUserEmail : session.account.username,
+                ShoppingList: [],
+                Users : [[session.account.username, session.account.name]]
+            })
+            console.log(newTrip)
+            await newTrip.save()
+            res.json({status:'success', ID: newTrip._id});
+        }else{
+            res.status(401).json({
+                tatus: "error",
+                error: "not logged in"
+            })
+        }   
     } catch (error) {
         res.status(500).send(error);
     }
@@ -33,22 +38,29 @@ router.post('/add', async function(req, res, next){
 router.post('/addUser', async function (req, res, next){
     let session = req.session
     try {
-        console.log("start")
-        console.log(req.body.tripID)
-        let trip = await req.models.Trip.findById(req.body.tripID)
-        console.log(trip);
-        let flag = true
-        for (let i = 0; i < trip.Users.length; i++) {
-            if (trip.Users[i].includes(session.account.username)) {
-                flag = false
-                break;
-            } 
+        if(req.session.isAuthenticated == true){
+            console.log("start")
+            console.log(req.body.tripID)
+            let trip = await req.models.Trip.findById(req.body.tripID)
+            console.log(trip);
+            let flag = true
+            for (let i = 0; i < trip.Users.length; i++) {
+                if (trip.Users[i].includes(session.account.username)) {
+                    flag = false
+                    break;
+                } 
+            }
+            if(flag){
+                trip.Users.push([session.account.username, session.account.name])
+            }
+            await trip.save()
+            res.json({status:'success'});
+        }else{
+            res.status(401).json({
+                tatus: "error",
+                error: "not logged in"
+            })
         }
-        if(flag){
-            trip.Users.push([session.account.username, session.account.name])
-        }
-        await trip.save()
-        res.json({status:'success'});
     } catch(error) {
         res.status(500).send(error);
     }
@@ -69,15 +81,25 @@ router.post('/addUser', async function (req, res, next){
 
 // this delete method removes a trip from the mongoDB database
 router.delete('/delete', async function(req, res, next){
+    let session = req.session
     try {
-        let tripID = req.query.tripID
-        //Only allow deletion if it's the primary user
-        //let trip = await req.models.trip.findById(tripID)
-        //if(trip.PrimaryUserEmail == session.account.username){
-            await req.models.Trip.deleteOne({_id:tripID})
-            res.json({status:'success'});
-        //}
+        if(req.session.isAuthenticated == true){
+            let tripID = req.query.tripID
+            //Only allow deletion if it's the primary user
+            let trip = await req.models.Trip.findById(tripID)
+            if(trip.PrimaryUserEmail === session.account.username){
+                await req.models.List.deleteMany({tripID : tripID})
+                await req.models.Trip.deleteOne({_id:tripID})
+                res.json({status:'success'});
+            }
+            }else{
+                res.status(401).json({
+                    tatus: "error",
+                    error: "not logged in"
+                })
+            }
     } catch (error) {
+        console.log(error)
         res.status(500).send(error);
     }
 })
